# app/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /products:
    put:
      tags:
      - products-controller
      operationId: editProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditProductDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
      - products-controller
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
    delete:
      tags:
      - products-controller
      operationId: deleteProduct
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /users/register:
    post:
      tags:
      - user-controller
      operationId: register
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /users/personal-information:
    post:
      tags:
      - user-controller
      operationId: addPersonalInformation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalInformation'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /users/login:
    post:
      tags:
      - user-controller
      operationId: login
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /system/personal-information:
    post:
      tags:
      - system-controller
      operationId: addPersonalInformation_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalInformation'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /employees/personal-information:
    post:
      tags:
      - employees-controller
      operationId: addPersonalInformation_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalInformation'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /discounts/personal-information:
    post:
      tags:
      - discounts-controller
      operationId: addPersonalInformation_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalInformation'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /clients/personal-information:
    post:
      tags:
      - clients-controller
      operationId: addPersonalInformation_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalInformation'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /products/{productId}/tax:
    get:
      tags:
      - products-controller
      operationId: getProductTaxes
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
    EditProductDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        quantity:
          type: integer
          format: int32
        price:
          type: number
        store:
          $ref: '#/components/schemas/Store'
    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        workHours:
          type: string
        address:
          type: string
        contactInfo:
          type: string
    PersonalInformation:
      type: object
      properties:
        fullName:
          type: string
        personalCode:
          type: string
        address:
          type: string
    CreateProductDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        quantity:
          type: integer
          format: int32
        price:
          type: number
        store:
          $ref: '#/components/schemas/Store'